18 search (#21)
* Update navigation.component.ts

* Data Existence Check Before Applying the Filter

* All cards matching the search term in any field will be displayed

Most Comprehensive Filter:
The applyFilter method now searches in all fields: nm_indicator, name_indicator, nm_completo_indicador, dc_conito_indicador and tx_fonte_indicador. This allows any term you type into the search bar to be found in any of these fields.

Conversion to Lower Case (toLowerCase):
To ensure that the search is not case sensitive, we use .toLowerCase() on both the filter value and the data fields.

shouldDisplayCard Method Updated:
We kept the shouldDisplayCard method to decide whether a card should be displayed, now based on filtered results.

* Show All Cards Matching Filter

If the user has not selected a strategy, the search bar will still show cards that match the search term.